+++ internal/application/dto/transaction.go
@@ -1,8 +1,8 @@
 package dto
 
 type TransactionRequest struct {
-	AccountId       int     `json:"account_id"`
-	OperationTypeId int     `json:"operation_type_id"`
+	AccountId       int64   `json:"account_id"`
+	OperationTypeId int64   `json:"operation_type_id"`
 	Amount          float64 `json:"amount"`
 }
