+++ internal/application/usecase/mocks/operation_type_repository_mock.go
@@ -3,9 +3,9 @@ package mocks
 import "context"
 
 type OperationTypeRepositoryMock struct {
-	ExistsByIDFunc func(ctx context.Context, operationTypeID int) (bool, error)
+	ExistsByIDFunc func(ctx context.Context, operationTypeID int64) (bool, error)
 }
 
-func (m *OperationTypeRepositoryMock) ExistsByID(ctx context.Context, operationTypeID int) (bool, error) {
+func (m *OperationTypeRepositoryMock) ExistsByID(ctx context.Context, operationTypeID int64) (bool, error) {
 	return m.ExistsByIDFunc(ctx, operationTypeID)
 }
