+++ internal/ports/account_repository.go
@@ -7,6 +7,6 @@ import (
 
 type AccountRepository interface {
 	Insert(ctx context.Context, documentHash string, documentEncrypted string) (*domain.Account, error)
-	GetById(ctx context.Context, accountId int) (*domain.Account, error)
-	ExistsById(ctx context.Context, accountId int) (bool, error)
+	GetById(ctx context.Context, accountId int64) (*domain.Account, error)
+	ExistsById(ctx context.Context, accountId int64) (bool, error)
 }
