+++ internal/ports/operation_type_repository.go
@@ -3,5 +3,5 @@ package ports
 import "context"
 
 type OperationTypeRepository interface {
-	ExistsByID(ctx context.Context, operationTypeID int) (bool, error)
+	ExistsByID(ctx context.Context, operationTypeID int64) (bool, error)
 }
